Begin the implementation of the ssol_draw function

+++ src/ssol_camera.h
@@ -38,7 +38,7 @@ struct ssol_camera {
 static FINLINE void
 camera_ray
   (const struct ssol_camera* cam,
-   const double sample[2], /* In [0, 1[ */
+   const float sample[2], /* In [0, 1[ */
    float org[3],
    float dir[3])
 {
@@ -48,8 +48,8 @@ camera_ray
   ASSERT(sample[0] >= 0 || sample[0] < 1.f);
   ASSERT(sample[1] >= 0 || sample[1] < 1.f);
 
-  f3_mulf(x, cam->axis_x, (float)sample[0]*2.f - 1.f);
-  f3_mulf(y, cam->axis_y, (float)sample[1]*2.f - 1.f);
+  f3_mulf(x, cam->axis_x, sample[0]*2.f - 1.f);
+  f3_mulf(y, cam->axis_y, sample[1]*2.f - 1.f);
   f3_add(dir, f3_add(dir, x, y), cam->axis_z);
   len = f3_normalize(dir, dir);
   ASSERT(len >= 1.e-6f);
